Clean up analog background regression test
