+++ lib/components/AnalogSimulationViewer.tsx
@@ -1,15 +1,15 @@
+import type { CircuitJson } from "circuit-json"
 import {
+  type ColorOverrides,
   convertCircuitJsonToSchematicSimulationSvg,
   convertCircuitJsonToSimulationGraphSvg,
-  type ColorOverrides,
 } from "circuit-to-svg"
-import { useEffect, useState, useMemo, useRef } from "react"
-import { useResizeHandling } from "../hooks/use-resize-handling"
-import { useMouseMatrixTransform } from "use-mouse-matrix-transform"
+import { useEffect, useMemo, useRef, useState } from "react"
 import { toString as transformToString } from "transformation-matrix"
-import type { CircuitJson } from "circuit-json"
-import { AnalogSimulationSelector } from "./AnalogSimulationSelector"
+import { useMouseMatrixTransform } from "use-mouse-matrix-transform"
+import { useResizeHandling } from "../hooks/use-resize-handling"
 import { getAnalogSimulationBackgroundColor } from "../utils/get-analog-simulation-background-color"
+import { AnalogSimulationSelector } from "./AnalogSimulationSelector"
 
 const DEFAULT_RENDER_WIDTH = 1200
 const DEFAULT_COMBINED_RENDER_ASPECT_RATIO = 1
